add schemaMapping for posts
