+++ src/main/java/org/example/fishtank/controller/PostResolver.java
@@ -4,6 +4,7 @@ import org.example.fishtank.model.dto.fishDto.ResponseFish;
 import org.example.fishtank.model.dto.postDto.ResponsePost;
 import org.example.fishtank.service.FishService;
 import org.example.fishtank.service.PostService;
+import org.springframework.graphql.data.method.annotation.Argument;
 import org.springframework.graphql.data.method.annotation.QueryMapping;
 import org.springframework.graphql.data.method.annotation.SchemaMapping;
 import org.springframework.stereotype.Controller;
@@ -28,9 +29,13 @@ public class PostResolver {
         return postService.getAllPost();
     }
 
-    @SchemaMapping(typeName = "ResponsePost", field = "fish")
-    public ResponseFish getPostsForFish(ResponsePost post) {
-        return fishService.findById(post.fishId());
+    @SchemaMapping(typeName = "ResponseFish", field = "posts")
+    public List<ResponsePost> getPostsForFish(ResponseFish fish) {
+        return postService.findByFishId(fish.id());
     }
 
+    @QueryMapping
+    public List<ResponsePost> getAllPostByFishId(@Argument Integer id) {
+        return postService.findByFishId(id);
+    }
 }
